+++ deepinv/physics/noise.py
@@ -721,3 +721,60 @@ class LogPoissonNoise(NoiseModel):
         N1_tilde = torch.poisson(self.N0 * torch.exp(-x * self.mu), generator=self.rng)
         y = -torch.log(N1_tilde / self.N0) / self.mu
         return y
+
+
+class SaltPepperNoise(NoiseModel):
+    r"""
+    SaltPepper noise :math:`y = \begin{cases} 0 & \text{if } z < p\\ x & \text{if } z \in [p, 1-s]\\ 1 & \text{if } z > 1 - s\end{cases}` with :math:`z\sim\mathcal{U}(0,1)`
+
+    This noise model is also known as impulse noise, is a form of noise sometimes seen on digital images.
+    For black-and-white or grayscale images, it presents as sparsely occurring white and black pixels,
+    giving the appearance of an image sprinkled with salt and pepper.
+
+    The parameters s and p control the amount of salt (pixel to 1) and pepper (pixel to 0) noise.
+
+    |sep|
+
+    :Examples:
+
+        Adding LogPoisson noise to a physics operator by setting the ``noise_model``
+        attribute of the physics operator:
+
+        >>> from deepinv.physics import Denoising, SaltPepperNoise
+        >>> import torch
+        >>> physics = Denoising()
+        >>> physics.noise_model = SaltPepperNoise()
+        >>> x = torch.rand(1, 1, 2, 2)
+        >>> y = physics(x)
+
+    :param float s: amount of salt noise.
+    :param float p: amount of pepper noise.
+    :param torch.Generator rng: (optional) a pseudorandom random number generator for the parameter generation.
+    """
+
+    def __init__(self, p=0.025, s=0.025, rng: torch.Generator = None):
+        super().__init__(rng=rng)
+        self.update_parameters(p=p, s=s)
+
+    def forward(self, x, p=None, s=None, seed: int = None, **kwargs):
+        r"""
+        Adds the noise to measurements x
+
+        :param torch.Tensor x: measurements
+        :param None, float, torch.Tensor s: amount of salt noise.
+            If not None, it will overwrite the current salt noise.
+        :param None, float, torch.Tensor p: amount of pepper noise.
+            If not None, it will overwrite the current pepper noise.
+        :param int seed: the seed for the random number generator, if `rng` is provided.
+        :returns: noisy measurements
+        """
+        self.update_parameters(p=self._float_to_tensor(p), s=self._float_to_tensor(s))
+        self.rng_manual_seed(seed)
+
+        proba_flip = self.s + self.p
+        proba_salt_vs_pepper = self.s / (self.s + self.p)
+
+        mask_flipped = (self.rand_like(x) < proba_flip).float()
+        mask_salt = (self.rand_like(x) < proba_salt_vs_pepper).float()
+        y = x * (1 - mask_flipped) + mask_flipped * mask_salt
+        return y